refactor: refactor core tests
