+++ cobra/test/test_core/test_gene.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+"""Test functions of gene.py"""
+
+
+def test_repr_html_(model):
+    assert '<table>' in model.genes[0]._repr_html_()
